moved progress bar to common components and created a index to export all common comps
