+++ src/components/common/index.ts
@@ -1,0 +1,5 @@
+export { default as ProgressBar } from './ProgressBar'
+export { default as Accordion } from './Accordion'
+export { default as NavigationArrows } from './NavigationArrows'
+export { default as PrivateRoute } from './PrivateRoute'
+export { default as ResponsiveStepWrapper } from './ResponsiveStepWrapper'
