+++ src/components/eligibility/Eligibility.tsx
@@ -3,7 +3,7 @@ import { Redirect, NavLink } from 'react-router-dom'
 import { Button, Typography } from '@material-ui/core'
 import { withRouter } from 'react-router-dom'
 import { useTranslation } from 'react-i18next'
-import ProgressBar from '../progressBar/ProgressBar'
+import ProgressBar from '../common/ProgressBar'
 import SageForm from '../form/SageForm'
 import { FORM_IDS } from '../form/types'
 import Separator from '../static/Separator'
